+++ db/migrate/20241019173940_create_associated_attrs.rb
@@ -3,7 +3,7 @@ class CreateAssociatedAttrs < ActiveRecord::Migration[7.1]
     create_table :associated_attrs do |t|
       t.string :attr_title
       t.json :associated_endpoints
-      t.json :general_info
+      t.string :general_info
 
       t.timestamps
     end